requirements-dev.txt
# Development Dependencies
# Core Django development tools
django-debug-toolbar>=6.0.0
django-extensions>=4.1.0

# Testing
pytest>=8.4.0
pytest-django>=4.11.0
factory-boy>=3.3.0
coverage>=7.0.0

# Code Quality
black>=25.9.0
flake8>=7.3.0
isort>=6.0.0

# API Development
djangorestframework>=3.15.0

# Additional development tools
ipython>=8.0.0
django-silk>=5.0.0  # For performance profiling
django-cors-headers>=4.0.0  # For CORS handling in development
